- ESPD-30 add selection technical professional criterion tests
git-svn-id: https://webgate.ec.europa.eu/CITnet/svn/ESPD/trunk@236 f47dfcff-13ca-4e7d-9efb-eacd05900603

+++ espd-web/src/main/java/eu/europa/ec/grow/espd/domain/BreachOfObligations.java
@@ -28,7 +28,7 @@ public class BreachOfObligations extends ExclusionCriterion {
     }
 
     @Override
-    public Integer getAmount() {
+    public Double getAmount() {
         throw new UnsupportedOperationException("Breach of obligations criterion does not have an amount.");
     }
