+++ espd-web/src/main/java/eu/europa/ec/grow/espd/domain/CriminalConvictions.java
@@ -27,7 +27,7 @@ public class CriminalConvictions extends ExclusionCriterion {
     }
 
     @Override
-    public Integer getAmount() {
+    public Double getAmount() {
         throw new UnsupportedOperationException("Criminal convictions criterion does not have an amount.");
     }
